DB410c: fix installer image
The GUI installer expects the images to be installed to be places on a VFAT
partition (so that images can be copy/pasted from any PC for example). So we
need to ensure that we mount this partition at boot.

Also disabling rootfs resize , since it's a waste of time in this case.

Change-Id: I1aae7b95c97d26ebd8f1341ca46a5774394e7cf1
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>